Lives && DeathWall Update and  DRY toegepast
