--- Vertical-Slice-SSB/Assets/Scripts/HealthScripts/JigglyPuffsLives.cs
@@ -1,28 +1,0 @@
-using TMPro;
-using UnityEngine;
-
-public class JigglyPuffsLives : MonoBehaviour
-{
-    public int startingLives = 3;
-    public int currentLives;
-    public WinScreen WinScreen;
-    public TextMeshProUGUI livesText;
-    private void Awake()
-    {
-        currentLives = startingLives;
-        livesText.text = startingLives.ToString() + " lives left";
-
-    }
-    public void DecreaseLives()
-    {
-        currentLives--;
-        livesText.text = currentLives.ToString() + " lives left";
-        if (currentLives <= 0)
-        {
-            gameObject.SetActive(false);
-            currentLives = 0;
-            WinScreen.JigglyPuffWin();
-        }
-
-    }
-}
