--- Vertical-Slice-SSB/Assets/Scripts/HealthScripts/JigglyPuffsLives.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c87ff5a9ae8023741a11ebc4b594fa81
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
