--- Vertical-Slice-SSB/Assets/Scripts/HealthScripts/KirbysLives.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 02806705c8d8d0748ad2c3b00f54a4fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
