+++ Vertical-Slice-SSB/Assets/Scripts/HealthScripts/Lives.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 053d7e951bc0eff4897f51945d2442c9
+guid: 7cfce516f6748ff4380e6bdd7acc77c7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
